Compile with no_keywords for glib compatibility (#2)
This compile switch makes Qt not define the shorter macros like
`signals`, `slots`, etc. One can still rely on the prefixed uppercase
ones, Q_SIGNALS, Q_SLOTS, Q_EMIT, etc.

This is a way to prevent the collision with `_GDBusInterfaceInfo`,
present in `frida-core.h` which uses a `signals` variable which collides
with the macro.

Ported the code to use the uglier but safer macros, except for the
`foreach` loops, which are converted to ranged for loops with `qAsConst`
to prevent possible detaches of shared containers.